refactor(web): reduce PatientDetailClient complexity by extracting hooks and components
Closes #933

Extract data-fetching logic into usePatientDetail, useAllergyForm, and
useAiSummary hooks. Move inline ConsentTab, InsuranceTab, and
HealthLogTab to separate component files. Main component reduced from
1271 to 817 lines (36% reduction) with focused, testable pieces.

Add 15 unit tests for the extracted hooks covering data fetching,
error handling, form state management, and AI summary generation.
