+++ apps/web/src/app/patients/[id]/PatientDetailClient.tsx
@@ -4,8 +4,7 @@ import { useState } from 'react';
 import dynamic from 'next/dynamic';
 import Link from 'next/link';
 import { useRouter } from 'next/navigation';
-import { useQuery, useQueryClient } from '@tanstack/react-query';
-import { type Patient, formatDate } from '@health-watchers/types';
+import { formatDate } from '@health-watchers/types';
 import {
   Badge,
   Button,
@@ -26,10 +25,18 @@ import {
   CreatePaymentIntentForm,
   type CreatePaymentData,
 } from '@/components/forms/CreatePaymentIntentForm';
-import { queryKeys } from '@/lib/queryKeys';
 import { API_V1 } from '@/lib/api';
 import { useAuth } from '@/context/AuthContext';
 import PhotoUpload from '@/components/patients/PhotoUpload';
+import {
+  usePatientDetail,
+  useAllergyForm,
+  useAiSummary,
+  type Allergy,
+} from '@/hooks/usePatientDetail';
+import { ConsentTab } from '@/components/patients/ConsentTab';
+import { InsuranceTab } from '@/components/patients/InsuranceTab';
+import { HealthLogTab } from '@/components/patients/HealthLogTab';
 
 const VitalSignsCharts = dynamic(() => import('@/components/patients/VitalSignsCharts'), {
   ssr: false,
@@ -58,50 +65,13 @@ const CommunicationTimeline = dynamic(
   { ssr: false }
 );
 
-interface EncounterResponse {
-  id: string;
-  patientId: string;
-  chiefComplaint: string;
-  status: string;
-  notes?: string;
-  diagnosis?: { code: string; description: string; isPrimary?: boolean }[];
-  vitalSigns?: Record<string, unknown>;
-  aiSummary?: string;
-  createdAt: string;
-}
-
-interface PaymentResponse {
-  id: string;
-  amount: string;
-  assetCode?: string;
-  status: string;
-  txHash?: string;
-  createdAt?: string;
-}
-
-interface Allergy {
-  _id: string;
-  allergen: string;
-  allergenType: string;
-  reaction: string;
-  severity: 'mild' | 'moderate' | 'severe' | 'life-threatening';
-  onsetDate?: string;
-  isActive: boolean;
-}
+// ── Utility functions ────────────────────────────────────────────────────────
 
 function severityVariant(severity: string) {
   if (severity === 'life-threatening') return 'danger';
   if (severity === 'severe') return 'danger';
   if (severity === 'moderate') return 'warning';
   return 'default';
-}
-
-const NETWORK = process.env.NEXT_PUBLIC_STELLAR_NETWORK ?? 'testnet';
-const EDIT_ROLES = new Set(['DOCTOR', 'CLINIC_ADMIN', 'SUPER_ADMIN']);
-
-function calcAge(dob: string): number {
-  const diff = Date.now() - new Date(dob).getTime();
-  return Math.floor(diff / (1000 * 60 * 60 * 24 * 365.25));
 }
 
 function statusVariant(status: string) {
@@ -117,6 +87,16 @@ function paymentVariant(status: string) {
   if (status === 'failed') return 'danger';
   return 'default';
 }
+
+function calcAge(dob: string): number {
+  const diff = Date.now() - new Date(dob).getTime();
+  return Math.floor(diff / (1000 * 60 * 60 * 24 * 365.25));
+}
+
+const NETWORK = process.env.NEXT_PUBLIC_STELLAR_NETWORK ?? 'testnet';
+const EDIT_ROLES = new Set(['DOCTOR', 'CLINIC_ADMIN', 'SUPER_ADMIN']);
+
+// ── Labels type ──────────────────────────────────────────────────────────────
 
 interface Labels {
   back: string;
@@ -146,6 +126,8 @@ interface Labels {
   systemId: string;
 }
 
+// ── Main component ───────────────────────────────────────────────────────────
+
 export default function PatientDetailClient({
   patientId,
   labels,
@@ -155,124 +137,42 @@ export default function PatientDetailClient({
 }) {
   const { user } = useAuth();
   const router = useRouter();
-  const queryClient = useQueryClient();
   const [tab, setTab] = useState('encounters');
   const [showPaymentForm, setShowPaymentForm] = useState(false);
-  const [aiSummary, setAiSummary] = useState<string | null>(null);
-  const [aiLoading, setAiLoading] = useState(false);
-  const [aiLastRun, setAiLastRun] = useState<Date | null>(null);
-  const [toast, setToast] = useState<{ message: string; type: 'success' | 'error' } | null>(null);
 
   const {
-    data: patient,
-    isLoading: patientLoading,
-    error: patientError,
-  } = useQuery<Patient>({
-    queryKey: queryKeys.patients.detail(patientId),
-    queryFn: async () => {
-      const res = await fetch(`${API_V1}/patients/${patientId}`);
-      if (res.status === 404) throw new Error('404');
-      if (!res.ok) throw new Error('Failed to load patient');
-      const data = await res.json();
-      return data.data;
-    },
-  });
+    patient,
+    patientLoading,
+    patientError,
+    encounters,
+    encountersLoading,
+    encountersError,
+    payments,
+    paymentsLoading,
+    paymentsError,
+    vitals,
+    vitalsLoading,
+    analytics,
+    analyticsLoading,
+    allergies,
+    allergiesLoading,
+    refetchAllergies,
+    invalidatePatient,
+    invalidateEncounters,
+    invalidatePayments,
+  } = usePatientDetail(patientId);
 
-  const {
-    data: encounters = [],
-    isLoading: encountersLoading,
-    error: encountersError,
-  } = useQuery<EncounterResponse[]>({
-    queryKey: queryKeys.encounters.byPatient(patientId),
-    queryFn: async () => {
-      const res = await fetch(`${API_V1}/encounters/patient/${patientId}`);
-      if (!res.ok) throw new Error('Failed to load encounters');
-      const data = await res.json();
-      return data.data ?? [];
-    },
-  });
-
-  const {
-    data: payments = [],
-    isLoading: paymentsLoading,
-    error: paymentsError,
-  } = useQuery<PaymentResponse[]>({
-    queryKey: queryKeys.payments.byPatient(patientId),
-    queryFn: async () => {
-      const res = await fetch(`${API_V1}/payments?patientId=${patientId}`);
-      if (!res.ok) throw new Error('Failed to load payments');
-      const data = await res.json();
-      return data.data ?? [];
-    },
-  });
-
-  const { data: vitals = [], isLoading: vitalsLoading } = useQuery({
-    queryKey: ['vitals', patientId],
-    queryFn: async () => {
-      const res = await fetch(`${API_V1}/patients/${patientId}/vitals`);
-      if (!res.ok) return [];
-      const data = await res.json();
-      return data.data ?? [];
-    },
-    staleTime: 5 * 60 * 1000, // cache 5 minutes
-  });
-
-  const { data: analytics, isLoading: analyticsLoading } = useQuery({
-    queryKey: ['analytics', patientId],
-    queryFn: async () => {
-      const res = await fetch(`${API_V1}/patients/${patientId}/analytics`);
-      if (!res.ok) return null;
-      const data = await res.json();
-      return data.data ?? null;
-    },
-    staleTime: 5 * 60 * 1000,
-  });
-
-  const {
-    data: allergies = [],
-    isLoading: allergiesLoading,
-    refetch: refetchAllergies,
-  } = useQuery<Allergy[]>({
-    queryKey: ['allergies', patientId],
-    queryFn: async () => {
-      const res = await fetch(`${API_V1}/patients/${patientId}/allergies`);
-      if (!res.ok) return [];
-      const data = await res.json();
-      return data.data ?? [];
-    },
-  });
-
-  const [showAllergyForm, setShowAllergyForm] = useState(false);
-  const [allergyForm, setAllergyForm] = useState({
-    allergen: '',
-    allergenType: 'drug',
-    reaction: '',
-    severity: 'mild',
-  });
-  const [allergySubmitting, setAllergySubmitting] = useState(false);
-
-  const handleAddAllergy = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setAllergySubmitting(true);
-    try {
-      const res = await fetch(`${API_V1}/patients/${patientId}/allergies`, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(allergyForm),
-      });
-      if (!res.ok) throw new Error('Failed to add allergy');
-      setShowAllergyForm(false);
-      setAllergyForm({ allergen: '', allergenType: 'drug', reaction: '', severity: 'mild' });
-      refetchAllergies();
-      setToast({ message: 'Allergy recorded.', type: 'success' });
-    } catch {
-      setToast({ message: 'Failed to add allergy.', type: 'error' });
-    } finally {
-      setAllergySubmitting(false);
-    }
-  };
+  const allergyState = useAllergyForm(patientId, refetchAllergies);
+  const aiState = useAiSummary(encounters);
 
   const canEdit = user && EDIT_ROLES.has(user.role);
+
+  // Merge toast state from allergy and AI hooks
+  const activeToast = allergyState.toast ?? aiState.toast;
+  const clearToast = () => {
+    allergyState.setToast(null);
+    aiState.setToast(null);
+  };
 
   const handleCreatePayment = async (data: CreatePaymentData) => {
     const res = await fetch(`${API_V1}/payments/intent`, {
@@ -285,32 +185,11 @@ export default function PatientDetailClient({
       throw new Error(body.message ?? `Error ${res.status}`);
     }
     setShowPaymentForm(false);
-    setToast({ message: 'Payment intent created.', type: 'success' });
-    queryClient.invalidateQueries({ queryKey: queryKeys.payments.byPatient(patientId) });
+    allergyState.setToast({ message: 'Payment intent created.', type: 'success' });
+    invalidatePayments();
   };
 
-  const handleGenerateAI = async () => {
-    if (!encounters.length) return;
-    setAiLoading(true);
-    try {
-      const res = await fetch(`${API_V1.replace('/api/v1', '')}/api/v1/ai/summarize`, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ encounterId: encounters[0].id }),
-      });
-      const body = await res.json();
-      if (!res.ok) throw new Error(body.message ?? 'AI unavailable');
-      setAiSummary(body.summary);
-      setAiLastRun(new Date());
-    } catch (err) {
-      setToast({
-        message: err instanceof Error ? err.message : 'AI generation failed',
-        type: 'error',
-      });
-    } finally {
-      setAiLoading(false);
-    }
-  };
+  // ── Loading / Error / Not Found states ───────────────────────────────────
 
   if (patientLoading) {
     return (
@@ -339,9 +218,7 @@ export default function PatientDetailClient({
       <PageWrapper className="py-8">
         <ErrorMessage
           message={patientError instanceof Error ? patientError.message : labels.error}
-          onRetry={() =>
-            queryClient.invalidateQueries({ queryKey: queryKeys.patients.detail(patientId) })
-          }
+          onRetry={invalidatePatient}
         />
       </PageWrapper>
     );
@@ -349,9 +226,13 @@ export default function PatientDetailClient({
 
   if (!patient) return null;
 
+  // ── Render ───────────────────────────────────────────────────────────────
+
   return (
     <PageWrapper className="py-8">
-      {toast && <Toast message={toast.message} type={toast.type} onClose={() => setToast(null)} />}
+      {activeToast && (
+        <Toast message={activeToast.message} type={activeToast.type} onClose={clearToast} />
+      )}
 
       {/* Breadcrumb */}
       <nav
@@ -372,83 +253,13 @@ export default function PatientDetailClient({
       </nav>
 
       {/* Demographics card */}
-      <section
-        aria-labelledby="demographics-heading"
-        className="mb-8 rounded-xl border border-neutral-200 bg-white p-6 shadow-sm"
-      >
-        <div className="mb-4 flex flex-wrap items-start justify-between gap-4">
-          <div className="flex items-start gap-4">
-            <PhotoUpload
-              patientId={patientId}
-              patientName={`${patient.firstName} ${patient.lastName}`}
-              photoUrl={(patient as any).photoUrl}
-              thumbnailUrl={(patient as any).thumbnailUrl}
-              canEdit={!!canEdit}
-            />
-            <div>
-              <h1 id="demographics-heading" className="text-2xl font-bold text-neutral-900">
-                {patient.firstName} {patient.lastName}
-              </h1>
-              <div className="mt-1 flex flex-wrap items-center gap-2">
-                <Badge variant={patient.gender === 'inactive' ? 'danger' : 'success'}>
-                  {labels.active}
-                </Badge>
-                <span className="text-xs text-neutral-500">
-                  {labels.registeredOn}: {formatDate((patient as any).createdAt)}
-                </span>
-              </div>
-            </div>
-          </div>
-          {canEdit && (
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={() => router.push(`/patients/${patientId}/edit`)}
-            >
-              {labels.editPatient}
-            </Button>
-          )}
-        </div>
-
-        <dl className="grid grid-cols-2 gap-x-6 gap-y-4 text-sm sm:grid-cols-3">
-          <div>
-            <dt className="text-xs font-semibold tracking-wide text-neutral-500 uppercase">
-              {labels.systemId}
-            </dt>
-            <dd className="mt-0.5 font-mono text-neutral-900">{patient.systemId}</dd>
-          </div>
-          <div>
-            <dt className="text-xs font-semibold tracking-wide text-neutral-500 uppercase">
-              {labels.dob}
-            </dt>
-            <dd className="mt-0.5 text-neutral-900">{formatDate(patient.dateOfBirth)}</dd>
-          </div>
-          <div>
-            <dt className="text-xs font-semibold tracking-wide text-neutral-500 uppercase">
-              {labels.age}
-            </dt>
-            <dd className="mt-0.5 text-neutral-900">{calcAge(patient.dateOfBirth)}</dd>
-          </div>
-          <div>
-            <dt className="text-xs font-semibold tracking-wide text-neutral-500 uppercase">
-              {labels.sex}
-            </dt>
-            <dd className="mt-0.5 text-neutral-900">{patient.sex}</dd>
-          </div>
-          <div>
-            <dt className="text-xs font-semibold tracking-wide text-neutral-500 uppercase">
-              {labels.contact}
-            </dt>
-            <dd className="mt-0.5 text-neutral-900">{patient.contactNumber || 'N/A'}</dd>
-          </div>
-          <div className="sm:col-span-2">
-            <dt className="text-xs font-semibold tracking-wide text-neutral-500 uppercase">
-              {labels.address}
-            </dt>
-            <dd className="mt-0.5 text-neutral-900">{patient.address || 'N/A'}</dd>
-          </div>
-        </dl>
-      </section>
+      <DemographicsCard
+        patient={patient}
+        patientId={patientId}
+        canEdit={!!canEdit}
+        labels={labels}
+        onEdit={() => router.push(`/patients/${patientId}/edit`)}
+      />
 
       {/* Tabs */}
       <Tabs value={tab} onValueChange={setTab}>
@@ -479,134 +290,34 @@ export default function PatientDetailClient({
           <TabsTrigger value="communications">Communications</TabsTrigger>
         </TabsList>
 
-        {/* Encounters tab */}
         <TabsContent value="encounters">
-          <div className="mb-4 flex items-center justify-between">
-            <p className="text-sm text-neutral-500">{encounters.length} record(s)</p>
-            <Button
-              size="sm"
-              variant="primary"
-              onClick={() => router.push(`/encounters/new?patientId=${patientId}`)}
-            >
-              + {labels.newEncounter}
-            </Button>
-          </div>
-
-          {encountersLoading ? (
-            <div className="space-y-3" aria-busy="true" aria-label={labels.loading}>
-              {[1, 2, 3].map((i) => (
-                <div key={i} className="h-20 animate-pulse rounded-lg bg-neutral-100" />
-              ))}
-            </div>
-          ) : encountersError ? (
-            <ErrorMessage
-              message={encountersError instanceof Error ? encountersError.message : labels.error}
-              onRetry={() =>
-                queryClient.invalidateQueries({
-                  queryKey: queryKeys.encounters.byPatient(patientId),
-                })
-              }
-            />
-          ) : encounters.length === 0 ? (
-            <EmptyState title={labels.noEncounters} icon="📋" />
-          ) : (
-            <ol className="space-y-3" aria-label={labels.encounters}>
-              {encounters.map((enc) => (
-                <li
-                  key={enc.id}
-                  className="rounded-lg border border-neutral-200 bg-white p-4 shadow-sm"
-                >
-                  <div className="flex flex-wrap items-start justify-between gap-2">
-                    <div>
-                      <p className="font-medium text-neutral-900">{enc.chiefComplaint}</p>
-                      <p className="mt-0.5 text-xs text-neutral-500">{formatDate(enc.createdAt)}</p>
-                    </div>
-                    <Badge variant={statusVariant(enc.status)}>{enc.status}</Badge>
-                  </div>
-                  {enc.notes && (
-                    <p className="mt-2 line-clamp-2 text-sm text-neutral-600">{enc.notes}</p>
-                  )}
-                  {enc.diagnosis && enc.diagnosis.length > 0 && (
-                    <p className="mt-1 text-xs text-neutral-500">
-                      Dx: {enc.diagnosis.map((d) => d.description).join(', ')}
-                    </p>
-                  )}
-                  {enc.aiSummary && (
-                    <details className="mt-2">
-                      <summary className="text-primary-600 cursor-pointer text-xs font-medium hover:underline">
-                        AI Summary
-                      </summary>
-                      <p className="mt-1 text-sm text-neutral-600">{enc.aiSummary}</p>
-                    </details>
-                  )}
-                </li>
-              ))}
-            </ol>
-          )}
+          <EncountersTabContent
+            encounters={encounters}
+            encountersLoading={encountersLoading}
+            encountersError={encountersError}
+            patientId={patientId}
+            labels={labels}
+            onRetry={invalidateEncounters}
+          />
         </TabsContent>
 
-        {/* Payments tab */}
         <TabsContent value="payments">
-          <div className="mb-4 flex items-center justify-between">
-            <p className="text-sm text-neutral-500">{payments.length} record(s)</p>
-            <Button size="sm" variant="primary" onClick={() => setShowPaymentForm(true)}>
-              + {labels.initiatePayment}
-            </Button>
-          </div>
-
-          {paymentsLoading ? (
-            <div className="space-y-3" aria-busy="true" aria-label={labels.loading}>
-              {[1, 2, 3].map((i) => (
-                <div key={i} className="h-16 animate-pulse rounded-lg bg-neutral-100" />
-              ))}
-            </div>
-          ) : paymentsError ? (
-            <ErrorMessage
-              message={paymentsError instanceof Error ? paymentsError.message : labels.error}
-              onRetry={() =>
-                queryClient.invalidateQueries({ queryKey: queryKeys.payments.byPatient(patientId) })
-              }
-            />
-          ) : payments.length === 0 ? (
-            <EmptyState title={labels.noPayments} icon="💳" />
-          ) : (
-            <ol className="space-y-3" aria-label={labels.payments}>
-              {payments.map((p) => (
-                <li
-                  key={p.id}
-                  className="rounded-lg border border-neutral-200 bg-white p-4 shadow-sm"
-                >
-                  <div className="flex flex-wrap items-start justify-between gap-2">
-                    <div>
-                      <p className="font-medium text-neutral-900">
-                        {p.amount}{' '}
-                        <span className="font-normal text-neutral-500">{p.assetCode ?? 'XLM'}</span>
-                      </p>
-                      <p className="mt-0.5 text-xs text-neutral-500">
-                        {p.createdAt ? new Date(p.createdAt).toLocaleDateString() : '—'}
-                      </p>
-                    </div>
-                    <Badge variant={paymentVariant(p.status)}>{p.status}</Badge>
-                  </div>
-                  {p.txHash && (
-                    <div className="mt-2">
-                      <StellarAddressDisplay value={p.txHash} type="tx" network={NETWORK} />
-                    </div>
-                  )}
-                </li>
-              ))}
-            </ol>
-          )}
+          <PaymentsTabContent
+            payments={payments}
+            paymentsLoading={paymentsLoading}
+            paymentsError={paymentsError}
+            labels={labels}
+            onInitiatePayment={() => setShowPaymentForm(true)}
+            onRetry={invalidatePayments}
+          />
         </TabsContent>
 
-        {/* Lab Results tab */}
         <TabsContent value="lab-results">
           <SectionErrorBoundary name="Lab Results">
             <LabResultsTab patientId={patientId} />
           </SectionErrorBoundary>
         </TabsContent>
 
-        {/* Vitals & Analytics tab */}
         <TabsContent value="vitals">
           <SectionErrorBoundary name="Vitals & Analytics">
             {vitalsLoading || analyticsLoading ? (
@@ -621,193 +332,53 @@ export default function PatientDetailClient({
           </SectionErrorBoundary>
         </TabsContent>
 
-        {/* Allergies tab */}
         <TabsContent value="allergies">
-          {canEdit && (
-            <div className="mb-4">
-              {showAllergyForm ? (
-                <form
-                  onSubmit={handleAddAllergy}
-                  className="space-y-3 rounded-lg border border-neutral-200 bg-white p-4 shadow-sm"
-                >
-                  <h3 className="font-medium text-neutral-900">Add Allergy</h3>
-                  <div className="grid grid-cols-2 gap-3">
-                    <input
-                      required
-                      placeholder="Allergen (e.g. Penicillin)"
-                      value={allergyForm.allergen}
-                      onChange={(e) => setAllergyForm((f) => ({ ...f, allergen: e.target.value }))}
-                      className="col-span-2 rounded border border-neutral-300 px-3 py-2 text-sm"
-                    />
-                    <select
-                      value={allergyForm.allergenType}
-                      onChange={(e) =>
-                        setAllergyForm((f) => ({ ...f, allergenType: e.target.value }))
-                      }
-                      className="rounded border border-neutral-300 px-3 py-2 text-sm"
-                    >
-                      <option value="drug">Drug</option>
-                      <option value="food">Food</option>
-                      <option value="environmental">Environmental</option>
-                      <option value="other">Other</option>
-                    </select>
-                    <select
-                      value={allergyForm.severity}
-                      onChange={(e) => setAllergyForm((f) => ({ ...f, severity: e.target.value }))}
-                      className="rounded border border-neutral-300 px-3 py-2 text-sm"
-                    >
-                      <option value="mild">Mild</option>
-                      <option value="moderate">Moderate</option>
-                      <option value="severe">Severe</option>
-                      <option value="life-threatening">Life-threatening</option>
-                    </select>
-                    <input
-                      required
-                      placeholder="Reaction (e.g. Anaphylaxis)"
-                      value={allergyForm.reaction}
-                      onChange={(e) => setAllergyForm((f) => ({ ...f, reaction: e.target.value }))}
-                      className="col-span-2 rounded border border-neutral-300 px-3 py-2 text-sm"
-                    />
-                  </div>
-                  <div className="flex gap-2">
-                    <button
-                      type="submit"
-                      disabled={allergySubmitting}
-                      className="bg-primary-600 hover:bg-primary-700 rounded px-4 py-2 text-sm text-white disabled:opacity-50"
-                    >
-                      {allergySubmitting ? 'Saving…' : 'Save Allergy'}
-                    </button>
-                    <button
-                      type="button"
-                      onClick={() => setShowAllergyForm(false)}
-                      className="rounded border border-neutral-300 px-4 py-2 text-sm text-neutral-700 hover:bg-neutral-50"
-                    >
-                      Cancel
-                    </button>
-                  </div>
-                </form>
-              ) : (
-                <Button size="sm" onClick={() => setShowAllergyForm(true)}>
-                  + Add Allergy
-                </Button>
-              )}
-            </div>
-          )}
-          {allergiesLoading ? (
-            <div className="space-y-3" aria-busy="true">
-              {[1, 2].map((i) => (
-                <div key={i} className="h-16 animate-pulse rounded-lg bg-neutral-100" />
-              ))}
-            </div>
-          ) : allergies.length === 0 ? (
-            <EmptyState title="No known allergies recorded" icon="💊" />
-          ) : (
-            <ol className="space-y-3" aria-label="Allergies">
-              {allergies.map((a) => (
-                <li
-                  key={a._id}
-                  className="rounded-lg border border-neutral-200 bg-white p-4 shadow-sm"
-                >
-                  <div className="flex flex-wrap items-start justify-between gap-2">
-                    <div>
-                      <p className="font-medium text-neutral-900">{a.allergen}</p>
-                      <p className="mt-0.5 text-xs text-neutral-500">
-                        {a.allergenType} · Reaction: {a.reaction}
-                        {a.onsetDate && ` · Onset: ${new Date(a.onsetDate).toLocaleDateString()}`}
-                      </p>
-                    </div>
-                    <Badge variant={severityVariant(a.severity)}>
-                      {a.severity === 'life-threatening' ? '⚠ ' : ''}
-                      {a.severity}
-                    </Badge>
-                  </div>
-                </li>
-              ))}
-            </ol>
-          )}
+          <AllergiesTabContent
+            allergies={allergies}
+            allergiesLoading={allergiesLoading}
+            canEdit={!!canEdit}
+            allergyState={allergyState}
+          />
         </TabsContent>
 
-        {/* AI Insights tab */}
         <TabsContent value="ai">
-          <div className="rounded-xl border border-blue-100 bg-white p-6 shadow-sm">
-            <div className="mb-4 flex items-center justify-between">
-              <div className="flex items-center gap-2">
-                <span className="rounded bg-blue-600 px-2 py-0.5 text-[10px] font-bold tracking-widest text-white">
-                  CLINICAL AI
-                </span>
-                <span className="text-sm font-semibold text-neutral-800">{labels.aiInsights}</span>
-              </div>
-              <Button
-                size="sm"
-                variant="outline"
-                onClick={handleGenerateAI}
-                disabled={aiLoading || encounters.length === 0}
-                aria-busy={aiLoading}
-              >
-                {aiLoading ? 'Generating…' : labels.generateSummary}
-              </Button>
-            </div>
-
-            {aiLoading ? (
-              <div className="space-y-2.5" aria-busy="true">
-                {[1, 2, 3, 4].map((i) => (
-                  <div
-                    key={i}
-                    className={`h-3.5 animate-pulse rounded bg-neutral-200 ${i === 1 ? 'w-[70%]' : i === 2 ? 'w-[77%]' : i === 3 ? 'w-[84%]' : 'w-[91%]'}`}
-                  />
-                ))}
-              </div>
-            ) : aiSummary ? (
-              <>
-                <p className="text-sm leading-relaxed text-neutral-600">{aiSummary}</p>
-                {aiLastRun && (
-                  <p className="mt-3 text-xs text-neutral-500">
-                    {labels.lastAnalysis}: {aiLastRun.toLocaleString()}
-                  </p>
-                )}
-              </>
-            ) : (
-              <p className="text-sm text-neutral-500">{labels.aiSummaryPlaceholder}</p>
-            )}
-          </div>
+          <AiInsightsTabContent aiState={aiState} encounters={encounters} labels={labels} />
         </TabsContent>
 
-        {/* Consent tab */}
         <TabsContent value="consent">
           <ConsentTab patientId={patientId} canEdit={!!canEdit} />
         </TabsContent>
-        {/* Risk tab */}
+
         <TabsContent value="risk">
           <SectionErrorBoundary name="Risk Assessment">
             <RiskTab patient={patient} patientId={patientId} apiV1={API_V1} />
           </SectionErrorBoundary>
         </TabsContent>
 
-        {/* Referrals tab */}
         <TabsContent value="referrals">
           <SectionErrorBoundary name="Referrals">
             <PatientReferralsTab patientId={patientId} />
           </SectionErrorBoundary>
         </TabsContent>
 
-        {/* Documents tab */}
         <TabsContent value="documents">
           <PatientDocumentsTab patientId={patientId} clinicId={user?.clinicId ?? ''} />
         </TabsContent>
 
-        {/* Care Plans tab */}
         <TabsContent value="care-plans">
           <CarePlanTab patientId={patientId} />
         </TabsContent>
 
-        {/* Health Log tab */}
         <TabsContent value="health-log">
           <HealthLogTab patientId={patientId} />
         </TabsContent>
 
-        {/* Communications tab */}
         <TabsContent value="communications">
           <CommunicationTimeline patientId={patientId} />
+        </TabsContent>
+
+        <TabsContent value="insurance">
+          <InsuranceTab patientId={patientId} canEdit={!!canEdit} />
         </TabsContent>
       </Tabs>
 
@@ -827,312 +398,322 @@ export default function PatientDetailClient({
   );
 }
 
-// ── Consent Tab (inline component) ───────────────────────────────────────────
-function ConsentTab({ patientId, canEdit }: { patientId: string; canEdit: boolean }) {
-  const queryClient = useQueryClient();
-  const [granting, setGranting] = useState<string | null>(null);
+// ── Sub-components ───────────────────────────────────────────────────────────
 
-  const CONSENT_TYPES = [
-    'treatment',
-    'data_sharing',
-    'ai_analysis',
-    'research',
-    'marketing',
-  ] as const;
-
-  const { data: consents = [], isLoading } = useQuery<any[]>({
-    queryKey: ['consents', patientId],
-    queryFn: async () => {
-      const res = await fetch(`${API_V1}/patients/${patientId}/consent`);
-      if (!res.ok) return [];
-      return (await res.json()).data ?? [];
-    },
-  });
-
-  const consentMap = Object.fromEntries(consents.map((c) => [c.type, c]));
-
-  async function grant(type: string) {
-    setGranting(type);
-    try {
-      await fetch(`${API_V1}/patients/${patientId}/consent`, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ type }),
-      });
-      queryClient.invalidateQueries({ queryKey: ['consents', patientId] });
-    } finally {
-      setGranting(null);
-    }
-  }
-
-  async function withdraw(type: string) {
-    setGranting(type);
-    try {
-      await fetch(`${API_V1}/patients/${patientId}/consent/${type}`, { method: 'DELETE' });
-      queryClient.invalidateQueries({ queryKey: ['consents', patientId] });
-    } finally {
-      setGranting(null);
-    }
-  }
-
-  if (isLoading)
-    return <div className="h-32 animate-pulse rounded bg-neutral-100" aria-busy="true" />;
-
+function DemographicsCard({
+  patient,
+  patientId,
+  canEdit,
+  labels,
+  onEdit,
+}: {
+  patient: any;
+  patientId: string;
+  canEdit: boolean;
+  labels: Labels;
+  onEdit: () => void;
+}) {
   return (
-    <section aria-label="Patient consent management">
-      <p className="mb-4 text-sm text-neutral-500">
-        Manage patient consent for data use and treatment. Withdrawal is immediate.
-      </p>
-      <ul className="space-y-3" role="list">
-        {CONSENT_TYPES.map((type) => {
-          const consent = consentMap[type];
-          const isGranted =
-            consent?.status === 'granted' &&
-            (!consent.expiresAt || new Date(consent.expiresAt) > new Date());
-          return (
-            <li
-              key={type}
-              className="flex items-center justify-between rounded-lg border border-neutral-200 bg-white p-4"
-            >
-              <div>
-                <p className="font-medium text-neutral-900 capitalize">{type.replace('_', ' ')}</p>
-                {consent?.grantedAt && (
-                  <p className="text-xs text-neutral-400">
-                    {isGranted ? 'Granted' : 'Withdrawn'}{' '}
-                    {new Date(consent.grantedAt).toLocaleDateString()}
-                    {consent.expiresAt &&
-                      ` · Expires ${new Date(consent.expiresAt).toLocaleDateString()}`}
-                  </p>
-                )}
-              </div>
-              <div className="flex items-center gap-2">
-                <span
-                  className={`inline-flex items-center rounded-full px-2 py-0.5 text-xs font-medium ${isGranted ? 'bg-green-100 text-green-700' : 'bg-neutral-100 text-neutral-500'}`}
-                >
-                  {isGranted
-                    ? 'Granted'
-                    : consent?.status === 'withdrawn'
-                      ? 'Withdrawn'
-                      : 'Not set'}
-                </span>
-                {canEdit &&
-                  (isGranted ? (
-                    <button
-                      onClick={() => withdraw(type)}
-                      disabled={granting === type}
-                      className="text-xs text-red-600 hover:underline focus:outline-none disabled:opacity-50"
-                      aria-label={`Withdraw ${type} consent`}
-                    >
-                      Withdraw
-                    </button>
-                  ) : (
-                    <button
-                      onClick={() => grant(type)}
-                      disabled={granting === type}
-                      className="text-primary-600 text-xs hover:underline focus:outline-none disabled:opacity-50"
-                      aria-label={`Grant ${type} consent`}
-                    >
-                      Grant
-                    </button>
-                  ))}
-              </div>
-            </li>
-          );
-        })}
-      </ul>
+    <section
+      aria-labelledby="demographics-heading"
+      className="mb-8 rounded-xl border border-neutral-200 bg-white p-6 shadow-sm"
+    >
+      <div className="mb-4 flex flex-wrap items-start justify-between gap-4">
+        <div className="flex items-start gap-4">
+          <PhotoUpload
+            patientId={patientId}
+            patientName={`${patient.firstName} ${patient.lastName}`}
+            photoUrl={patient.photoUrl}
+            thumbnailUrl={patient.thumbnailUrl}
+            canEdit={canEdit}
+          />
+          <div>
+            <h1 id="demographics-heading" className="text-2xl font-bold text-neutral-900">
+              {patient.firstName} {patient.lastName}
+            </h1>
+            <div className="mt-1 flex flex-wrap items-center gap-2">
+              <Badge variant={patient.gender === 'inactive' ? 'danger' : 'success'}>
+                {labels.active}
+              </Badge>
+              <span className="text-xs text-neutral-500">
+                {labels.registeredOn}: {formatDate(patient.createdAt)}
+              </span>
+            </div>
+          </div>
+        </div>
+        {canEdit && (
+          <Button variant="outline" size="sm" onClick={onEdit}>
+            {labels.editPatient}
+          </Button>
+        )}
+      </div>
+
+      <dl className="grid grid-cols-2 gap-x-6 gap-y-4 text-sm sm:grid-cols-3">
+        <div>
+          <dt className="text-xs font-semibold tracking-wide text-neutral-500 uppercase">
+            {labels.systemId}
+          </dt>
+          <dd className="mt-0.5 font-mono text-neutral-900">{patient.systemId}</dd>
+        </div>
+        <div>
+          <dt className="text-xs font-semibold tracking-wide text-neutral-500 uppercase">
+            {labels.dob}
+          </dt>
+          <dd className="mt-0.5 text-neutral-900">{formatDate(patient.dateOfBirth)}</dd>
+        </div>
+        <div>
+          <dt className="text-xs font-semibold tracking-wide text-neutral-500 uppercase">
+            {labels.age}
+          </dt>
+          <dd className="mt-0.5 text-neutral-900">{calcAge(patient.dateOfBirth)}</dd>
+        </div>
+        <div>
+          <dt className="text-xs font-semibold tracking-wide text-neutral-500 uppercase">
+            {labels.sex}
+          </dt>
+          <dd className="mt-0.5 text-neutral-900">{patient.sex}</dd>
+        </div>
+        <div>
+          <dt className="text-xs font-semibold tracking-wide text-neutral-500 uppercase">
+            {labels.contact}
+          </dt>
+          <dd className="mt-0.5 text-neutral-900">{patient.contactNumber || 'N/A'}</dd>
+        </div>
+        <div className="sm:col-span-2">
+          <dt className="text-xs font-semibold tracking-wide text-neutral-500 uppercase">
+            {labels.address}
+          </dt>
+          <dd className="mt-0.5 text-neutral-900">{patient.address || 'N/A'}</dd>
+        </div>
+      </dl>
     </section>
   );
 }
 
-// ── Insurance Tab (inline component) ─────────────────────────────────────────
-interface InsuranceRecord {
-  _id: string;
-  provider: string;
-  policyNumber: string;
-  groupNumber?: string;
-  coverageType: string;
-  effectiveDate?: string;
-  expirationDate?: string;
-  isPrimary: boolean;
-}
-
-const COVERAGE_TYPES = [
-  'HMO',
-  'PPO',
-  'EPO',
-  'POS',
-  'HDHP',
-  'Medicare',
-  'Medicaid',
-  'other',
-] as const;
-
-const EMPTY_INSURANCE = {
-  provider: '',
-  policyNumber: '',
-  groupNumber: '',
-  coverageType: 'PPO',
-  effectiveDate: '',
-  expirationDate: '',
-  isPrimary: false,
-};
-
-function InsuranceTab({ patientId, canEdit }: { patientId: string; canEdit: boolean }) {
-  const queryClient = useQueryClient();
-  const [showForm, setShowForm] = useState(false);
-  const [submitting, setSubmitting] = useState(false);
-  const [form, setForm] = useState({ ...EMPTY_INSURANCE });
-  const [error, setError] = useState<string | null>(null);
-
-  const queryKey = ['insurance', patientId];
-
-  const { data: records = [], isLoading } = useQuery<InsuranceRecord[]>({
-    queryKey,
-    queryFn: async () => {
-      const res = await fetch(`${API_V1}/patients/${patientId}/insurance`);
-      if (!res.ok) return [];
-      return (await res.json()).data ?? [];
-    },
-  });
-
-  const refresh = () => queryClient.invalidateQueries({ queryKey });
-
-  async function handleAdd(e: React.FormEvent) {
-    e.preventDefault();
-    setSubmitting(true);
-    setError(null);
-    // Strip empty optional fields so the API does not reject empty-string dates.
-    const payload: Record<string, unknown> = {
-      provider: form.provider,
-      policyNumber: form.policyNumber,
-      coverageType: form.coverageType,
-      isPrimary: form.isPrimary,
-    };
-    if (form.groupNumber) payload.groupNumber = form.groupNumber;
-    if (form.effectiveDate) payload.effectiveDate = form.effectiveDate;
-    if (form.expirationDate) payload.expirationDate = form.expirationDate;
-    try {
-      const res = await fetch(`${API_V1}/patients/${patientId}/insurance`, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(payload),
-      });
-      if (!res.ok) {
-        const body = await res.json().catch(() => ({}));
-        throw new Error(body.message ?? 'Failed to add insurance');
-      }
-      setShowForm(false);
-      setForm({ ...EMPTY_INSURANCE });
-      refresh();
-    } catch (err) {
-      setError(err instanceof Error ? err.message : 'Failed to add insurance');
-    } finally {
-      setSubmitting(false);
-    }
-  }
-
-  async function setPrimary(record: InsuranceRecord) {
-    await fetch(`${API_V1}/patients/${patientId}/insurance/${record._id}`, {
-      method: 'PATCH',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ isPrimary: true }),
-    });
-    refresh();
-  }
-
-  async function remove(record: InsuranceRecord) {
-    await fetch(`${API_V1}/patients/${patientId}/insurance/${record._id}`, { method: 'DELETE' });
-    refresh();
-  }
-
-  if (isLoading)
-    return <div className="h-32 animate-pulse rounded bg-neutral-100" aria-busy="true" />;
+function EncountersTabContent({
+  encounters,
+  encountersLoading,
+  encountersError,
+  patientId,
+  labels,
+  onRetry,
+}: {
+  encounters: any[];
+  encountersLoading: boolean;
+  encountersError: Error | null;
+  patientId: string;
+  labels: Labels;
+  onRetry: () => void;
+}) {
+  const router = useRouter();
 
   return (
-    <section aria-label="Patient insurance management">
+    <>
+      <div className="mb-4 flex items-center justify-between">
+        <p className="text-sm text-neutral-500">{encounters.length} record(s)</p>
+        <Button
+          size="sm"
+          variant="primary"
+          onClick={() => router.push(`/encounters/new?patientId=${patientId}`)}
+        >
+          + {labels.newEncounter}
+        </Button>
+      </div>
+
+      {encountersLoading ? (
+        <div className="space-y-3" aria-busy="true" aria-label={labels.loading}>
+          {[1, 2, 3].map((i) => (
+            <div key={i} className="h-20 animate-pulse rounded-lg bg-neutral-100" />
+          ))}
+        </div>
+      ) : encountersError ? (
+        <ErrorMessage
+          message={encountersError instanceof Error ? encountersError.message : labels.error}
+          onRetry={onRetry}
+        />
+      ) : encounters.length === 0 ? (
+        <EmptyState title={labels.noEncounters} icon="📋" />
+      ) : (
+        <ol className="space-y-3" aria-label={labels.encounters}>
+          {encounters.map((enc) => (
+            <li
+              key={enc.id}
+              className="rounded-lg border border-neutral-200 bg-white p-4 shadow-sm"
+            >
+              <div className="flex flex-wrap items-start justify-between gap-2">
+                <div>
+                  <p className="font-medium text-neutral-900">{enc.chiefComplaint}</p>
+                  <p className="mt-0.5 text-xs text-neutral-500">{formatDate(enc.createdAt)}</p>
+                </div>
+                <Badge variant={statusVariant(enc.status)}>{enc.status}</Badge>
+              </div>
+              {enc.notes && (
+                <p className="mt-2 line-clamp-2 text-sm text-neutral-600">{enc.notes}</p>
+              )}
+              {enc.diagnosis && enc.diagnosis.length > 0 && (
+                <p className="mt-1 text-xs text-neutral-500">
+                  Dx: {enc.diagnosis.map((d: any) => d.description).join(', ')}
+                </p>
+              )}
+              {enc.aiSummary && (
+                <details className="mt-2">
+                  <summary className="text-primary-600 cursor-pointer text-xs font-medium hover:underline">
+                    AI Summary
+                  </summary>
+                  <p className="mt-1 text-sm text-neutral-600">{enc.aiSummary}</p>
+                </details>
+              )}
+            </li>
+          ))}
+        </ol>
+      )}
+    </>
+  );
+}
+
+function PaymentsTabContent({
+  payments,
+  paymentsLoading,
+  paymentsError,
+  labels,
+  onInitiatePayment,
+  onRetry,
+}: {
+  payments: any[];
+  paymentsLoading: boolean;
+  paymentsError: Error | null;
+  labels: Labels;
+  onInitiatePayment: () => void;
+  onRetry: () => void;
+}) {
+  return (
+    <>
+      <div className="mb-4 flex items-center justify-between">
+        <p className="text-sm text-neutral-500">{payments.length} record(s)</p>
+        <Button size="sm" variant="primary" onClick={onInitiatePayment}>
+          + {labels.initiatePayment}
+        </Button>
+      </div>
+
+      {paymentsLoading ? (
+        <div className="space-y-3" aria-busy="true" aria-label={labels.loading}>
+          {[1, 2, 3].map((i) => (
+            <div key={i} className="h-16 animate-pulse rounded-lg bg-neutral-100" />
+          ))}
+        </div>
+      ) : paymentsError ? (
+        <ErrorMessage
+          message={paymentsError instanceof Error ? paymentsError.message : labels.error}
+          onRetry={onRetry}
+        />
+      ) : payments.length === 0 ? (
+        <EmptyState title={labels.noPayments} icon="💳" />
+      ) : (
+        <ol className="space-y-3" aria-label={labels.payments}>
+          {payments.map((p) => (
+            <li key={p.id} className="rounded-lg border border-neutral-200 bg-white p-4 shadow-sm">
+              <div className="flex flex-wrap items-start justify-between gap-2">
+                <div>
+                  <p className="font-medium text-neutral-900">
+                    {p.amount}{' '}
+                    <span className="font-normal text-neutral-500">{p.assetCode ?? 'XLM'}</span>
+                  </p>
+                  <p className="mt-0.5 text-xs text-neutral-500">
+                    {p.createdAt ? new Date(p.createdAt).toLocaleDateString() : '—'}
+                  </p>
+                </div>
+                <Badge variant={paymentVariant(p.status)}>{p.status}</Badge>
+              </div>
+              {p.txHash && (
+                <div className="mt-2">
+                  <StellarAddressDisplay value={p.txHash} type="tx" network={NETWORK} />
+                </div>
+              )}
+            </li>
+          ))}
+        </ol>
+      )}
+    </>
+  );
+}
+
+function AllergiesTabContent({
+  allergies,
+  allergiesLoading,
+  canEdit,
+  allergyState,
+}: {
+  allergies: Allergy[];
+  allergiesLoading: boolean;
+  canEdit: boolean;
+  allergyState: ReturnType<typeof useAllergyForm>;
+}) {
+  const {
+    showAllergyForm,
+    setShowAllergyForm,
+    allergyForm,
+    setAllergyForm,
+    allergySubmitting,
+    handleAddAllergy,
+  } = allergyState;
+
+  return (
+    <>
       {canEdit && (
         <div className="mb-4">
-          {showForm ? (
+          {showAllergyForm ? (
             <form
-              onSubmit={handleAdd}
+              onSubmit={handleAddAllergy}
               className="space-y-3 rounded-lg border border-neutral-200 bg-white p-4 shadow-sm"
             >
-              <h3 className="font-medium text-neutral-900">Add Insurance</h3>
-              {error && <p className="text-danger-600 text-sm">{error}</p>}
+              <h3 className="font-medium text-neutral-900">Add Allergy</h3>
               <div className="grid grid-cols-2 gap-3">
                 <input
                   required
-                  placeholder="Provider (e.g. Aetna)"
-                  value={form.provider}
-                  onChange={(e) => setForm((f) => ({ ...f, provider: e.target.value }))}
+                  placeholder="Allergen (e.g. Penicillin)"
+                  value={allergyForm.allergen}
+                  onChange={(e) => setAllergyForm((f) => ({ ...f, allergen: e.target.value }))}
                   className="col-span-2 rounded border border-neutral-300 px-3 py-2 text-sm"
                 />
+                <select
+                  value={allergyForm.allergenType}
+                  onChange={(e) => setAllergyForm((f) => ({ ...f, allergenType: e.target.value }))}
+                  className="rounded border border-neutral-300 px-3 py-2 text-sm"
+                >
+                  <option value="drug">Drug</option>
+                  <option value="food">Food</option>
+                  <option value="environmental">Environmental</option>
+                  <option value="other">Other</option>
+                </select>
+                <select
+                  value={allergyForm.severity}
+                  onChange={(e) => setAllergyForm((f) => ({ ...f, severity: e.target.value }))}
+                  className="rounded border border-neutral-300 px-3 py-2 text-sm"
+                >
+                  <option value="mild">Mild</option>
+                  <option value="moderate">Moderate</option>
+                  <option value="severe">Severe</option>
+                  <option value="life-threatening">Life-threatening</option>
+                </select>
                 <input
                   required
-                  placeholder="Policy Number"
-                  value={form.policyNumber}
-                  onChange={(e) => setForm((f) => ({ ...f, policyNumber: e.target.value }))}
-                  className="rounded border border-neutral-300 px-3 py-2 text-sm"
+                  placeholder="Reaction (e.g. Anaphylaxis)"
+                  value={allergyForm.reaction}
+                  onChange={(e) => setAllergyForm((f) => ({ ...f, reaction: e.target.value }))}
+                  className="col-span-2 rounded border border-neutral-300 px-3 py-2 text-sm"
                 />
-                <input
-                  placeholder="Group Number"
-                  value={form.groupNumber}
-                  onChange={(e) => setForm((f) => ({ ...f, groupNumber: e.target.value }))}
-                  className="rounded border border-neutral-300 px-3 py-2 text-sm"
-                />
-                <select
-                  value={form.coverageType}
-                  onChange={(e) => setForm((f) => ({ ...f, coverageType: e.target.value }))}
-                  className="rounded border border-neutral-300 px-3 py-2 text-sm"
-                  aria-label="Coverage type"
-                >
-                  {COVERAGE_TYPES.map((t) => (
-                    <option key={t} value={t}>
-                      {t}
-                    </option>
-                  ))}
-                </select>
-                <label className="flex items-center gap-2 text-sm text-neutral-700">
-                  <input
-                    type="checkbox"
-                    checked={form.isPrimary}
-                    onChange={(e) => setForm((f) => ({ ...f, isPrimary: e.target.checked }))}
-                  />
-                  Primary insurance
-                </label>
-                <label className="text-xs text-neutral-500">
-                  Effective date
-                  <input
-                    type="date"
-                    value={form.effectiveDate}
-                    onChange={(e) => setForm((f) => ({ ...f, effectiveDate: e.target.value }))}
-                    className="mt-0.5 block w-full rounded border border-neutral-300 px-3 py-2 text-sm text-neutral-900"
-                  />
-                </label>
-                <label className="text-xs text-neutral-500">
-                  Expiration date
-                  <input
-                    type="date"
-                    value={form.expirationDate}
-                    onChange={(e) => setForm((f) => ({ ...f, expirationDate: e.target.value }))}
-                    className="mt-0.5 block w-full rounded border border-neutral-300 px-3 py-2 text-sm text-neutral-900"
-                  />
-                </label>
               </div>
               <div className="flex gap-2">
                 <button
                   type="submit"
-                  disabled={submitting}
+                  disabled={allergySubmitting}
                   className="bg-primary-600 hover:bg-primary-700 rounded px-4 py-2 text-sm text-white disabled:opacity-50"
                 >
-                  {submitting ? 'Saving…' : 'Save Insurance'}
+                  {allergySubmitting ? 'Saving...' : 'Save Allergy'}
                 </button>
                 <button
                   type="button"
-                  onClick={() => {
-                    setShowForm(false);
-                    setError(null);
-                  }}
+                  onClick={() => setShowAllergyForm(false)}
                   className="rounded border border-neutral-300 px-4 py-2 text-sm text-neutral-700 hover:bg-neutral-50"
                 >
                   Cancel
@@ -1140,131 +721,97 @@ function InsuranceTab({ patientId, canEdit }: { patientId: string; canEdit: bool
               </div>
             </form>
           ) : (
-            <Button size="sm" onClick={() => setShowForm(true)}>
-              + Add Insurance
+            <Button size="sm" onClick={() => setShowAllergyForm(true)}>
+              + Add Allergy
             </Button>
           )}
         </div>
       )}
-
-      {records.length === 0 ? (
-        <EmptyState title="No insurance information on file" icon="🛡️" />
+      {allergiesLoading ? (
+        <div className="space-y-3" aria-busy="true">
+          {[1, 2].map((i) => (
+            <div key={i} className="h-16 animate-pulse rounded-lg bg-neutral-100" />
+          ))}
+        </div>
+      ) : allergies.length === 0 ? (
+        <EmptyState title="No known allergies recorded" icon="💊" />
       ) : (
-        <ol className="space-y-3" aria-label="Insurance records">
-          {records.map((ins) => (
-            <li
-              key={ins._id}
-              className="rounded-lg border border-neutral-200 bg-white p-4 shadow-sm"
-            >
+        <ol className="space-y-3" aria-label="Allergies">
+          {allergies.map((a) => (
+            <li key={a._id} className="rounded-lg border border-neutral-200 bg-white p-4 shadow-sm">
               <div className="flex flex-wrap items-start justify-between gap-2">
                 <div>
-                  <p className="flex items-center gap-2 font-medium text-neutral-900">
-                    {ins.provider}
-                    {ins.isPrimary && <Badge variant="primary">Primary</Badge>}
-                  </p>
+                  <p className="font-medium text-neutral-900">{a.allergen}</p>
                   <p className="mt-0.5 text-xs text-neutral-500">
-                    {ins.coverageType} · Policy {ins.policyNumber}
-                    {ins.groupNumber && ` · Group ${ins.groupNumber}`}
+                    {a.allergenType} · Reaction: {a.reaction}
+                    {a.onsetDate && ` · Onset: ${new Date(a.onsetDate).toLocaleDateString()}`}
                   </p>
-                  {(ins.effectiveDate || ins.expirationDate) && (
-                    <p className="mt-0.5 text-xs text-neutral-400">
-                      {ins.effectiveDate ? `Effective ${formatDate(ins.effectiveDate)}` : ''}
-                      {ins.expirationDate ? ` · Expires ${formatDate(ins.expirationDate)}` : ''}
-                    </p>
-                  )}
                 </div>
-                {canEdit && (
-                  <div className="flex items-center gap-3">
-                    {!ins.isPrimary && (
-                      <button
-                        onClick={() => setPrimary(ins)}
-                        className="text-primary-600 text-xs hover:underline focus:outline-none"
-                        aria-label={`Set ${ins.provider} as primary insurance`}
-                      >
-                        Set primary
-                      </button>
-                    )}
-                    <button
-                      onClick={() => remove(ins)}
-                      className="text-xs text-red-600 hover:underline focus:outline-none"
-                      aria-label={`Delete ${ins.provider} insurance`}
-                    >
-                      Delete
-                    </button>
-                  </div>
-                )}
+                <Badge variant={severityVariant(a.severity)}>
+                  {a.severity === 'life-threatening' ? '⚠ ' : ''}
+                  {a.severity}
+                </Badge>
               </div>
             </li>
           ))}
         </ol>
       )}
-    </section>
+    </>
   );
 }
 
-// ── Health Log Tab (inline component) ────────────────────────────────────────
-function HealthLogTab({ patientId }: { patientId: string }) {
-  const { data, isLoading, error } = useQuery<
-    {
-      metricType: string;
-      value: number;
-      unit: string;
-      loggedAt: string;
-      notes?: string;
-      flagged: boolean;
-    }[]
-  >({
-    queryKey: ['health-log', patientId],
-    queryFn: async () => {
-      const res = await fetch(`${API_V1}/patients/${patientId}/health-log`);
-      if (!res.ok) throw new Error('Failed to load health log');
-      return (await res.json()).data ?? [];
-    },
-  });
-
-  if (isLoading) return <div className="h-24 animate-pulse rounded bg-neutral-100" />;
-  if (error) return <ErrorMessage message="Failed to load health log" />;
-  if (!data?.length) return <EmptyState title="No health metrics logged yet" icon="📊" />;
+function AiInsightsTabContent({
+  aiState,
+  encounters,
+  labels,
+}: {
+  aiState: ReturnType<typeof useAiSummary>;
+  encounters: any[];
+  labels: Labels;
+}) {
+  const { aiSummary, aiLoading, aiLastRun, handleGenerateAI } = aiState;
 
   return (
-    <section aria-label="Patient health log">
-      <p className="mb-3 text-sm text-neutral-500">{data.length} entries</p>
-      <div className="overflow-x-auto">
-        <table className="w-full text-left text-sm">
-          <thead>
-            <tr className="border-b border-neutral-100 text-xs font-medium tracking-wide text-neutral-500 uppercase">
-              <th className="pr-4 pb-2">Date</th>
-              <th className="pr-4 pb-2">Metric</th>
-              <th className="pr-4 pb-2">Value</th>
-              <th className="pr-4 pb-2">Notes</th>
-              <th className="pb-2">Status</th>
-            </tr>
-          </thead>
-          <tbody>
-            {data.map((l, i) => (
-              <tr key={i} className="border-b border-neutral-50 hover:bg-neutral-50">
-                <td className="py-2 pr-4 text-neutral-500">
-                  {new Date(l.loggedAt).toLocaleString()}
-                </td>
-                <td className="py-2 pr-4 text-neutral-700 capitalize">
-                  {l.metricType.replace('_', ' ')}
-                </td>
-                <td className="py-2 pr-4 font-medium text-neutral-900">
-                  {l.value} {l.unit}
-                </td>
-                <td className="py-2 pr-4 text-neutral-500">{l.notes ?? '—'}</td>
-                <td className="py-2">
-                  {l.flagged ? (
-                    <Badge variant="danger">⚠ Abnormal</Badge>
-                  ) : (
-                    <Badge variant="success">Normal</Badge>
-                  )}
-                </td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
+    <div className="rounded-xl border border-blue-100 bg-white p-6 shadow-sm">
+      <div className="mb-4 flex items-center justify-between">
+        <div className="flex items-center gap-2">
+          <span className="rounded bg-blue-600 px-2 py-0.5 text-[10px] font-bold tracking-widest text-white">
+            CLINICAL AI
+          </span>
+          <span className="text-sm font-semibold text-neutral-800">{labels.aiInsights}</span>
+        </div>
+        <Button
+          size="sm"
+          variant="outline"
+          onClick={handleGenerateAI}
+          disabled={aiLoading || encounters.length === 0}
+          aria-busy={aiLoading}
+        >
+          {aiLoading ? 'Generating...' : labels.generateSummary}
+        </Button>
       </div>
-    </section>
+
+      {aiLoading ? (
+        <div className="space-y-2.5" aria-busy="true">
+          {[1, 2, 3, 4].map((i) => (
+            <div
+              key={i}
+              className={`h-3.5 animate-pulse rounded bg-neutral-200 ${i === 1 ? 'w-[70%]' : i === 2 ? 'w-[77%]' : i === 3 ? 'w-[84%]' : 'w-[91%]'}`}
+            />
+          ))}
+        </div>
+      ) : aiSummary ? (
+        <>
+          <p className="text-sm leading-relaxed text-neutral-600">{aiSummary}</p>
+          {aiLastRun && (
+            <p className="mt-3 text-xs text-neutral-500">
+              {labels.lastAnalysis}: {aiLastRun.toLocaleString()}
+            </p>
+          )}
+        </>
+      ) : (
+        <p className="text-sm text-neutral-500">{labels.aiSummaryPlaceholder}</p>
+      )}
+    </div>
   );
 }
